+++ src/data/clients/storage/DataStoreStorageClient.ts
@@ -2,9 +2,9 @@ import { D2Api, DataStore } from "../../../types/d2-api";
 import { Instance } from "../../entities/Instance";
 import { getD2APiFromInstance } from "../../utils/d2-api";
 import { dataStoreNamespace } from "./Namespaces";
-import { StorageDataSource } from "./StorageDataSource";
+import { StorageClient } from "./StorageClient";
 
-export class DataStoreStorageDataSource extends StorageDataSource {
+export class DataStoreStorageClient extends StorageClient {
     private api: D2Api;
     private dataStore: DataStore;
